MFDecoupling.jl Documentation

MFDecoupling.jl Documentation

Index

List of Functions

MFDecoupling

Index

MFDecoupling.gen_helpers

MFDecoupling.read_inputs

MFDecoupling.rhs!

MFDecoupling.rhs_real!

MFDecoupling.setup_calculation

List of Functions

MFDecoupling

MFDecoupling.gen_helpers — Method

gen_helpers(L::Int)

Generates index helpers for right hand side.

MFDecoupling.read_inputs — Method

read_inputs(fp1::String, fp2::String, L::Int)

Reads two files at locations fp1, fp2.... ?

MFDecoupling.rhs! — Method

rhs!(dX::Vector, X::Vector, p, t)::Nothing

RHS for DGL.

Parameters

p[1]: L, size of the chain

p[2]: U, local Hubbard interaction on the impurity site

p[3]: Vr, coupling between the impurity and chain

p[4]: μ_imp chemical potential for impurity

p[5]: - μ_c chemical potential for chain (onsite energy shift is includid in the chemical potential)

MFDecoupling.rhs_real! — Method

rhs!(dX::Vector, X::Vector, p, t)::Nothing

RHS for DGL.

Parameters

p[1]: L, size of the chain

p[2]: U, local Hubbard interaction on the impurity site

p[3]: Vr, coupling between the impurity and chain

p[4]: μ_imp chemical potential for impurity

p[5]: - μ_c chemical potential for chain (onsite energy shift is includid in the chemical potential)

LC = (L+3)L/2 LK = (L-1)L/2 LIm= 10 + LC +LK

Q[i,j] = (L+1)i - (i+1)i/2 + j i <= j (i=j= 0 is not considered) 1, ..., LC P[i,j] = (i-1)L - (i+1)i/2 + j 0 < i < j 1, ..., LK

Matrix Elelements ρ: [1,1], [1,2], [1,3], [1,4], [2,2], [2,3], [2,4], [3,3], [3,4], [4,4] Re -> [1, ..., 10] and Im -> [11 + LC + LK, ..., 20 + LC + LK]

Re(ρ[1,1]) -> X[1],             Re(ρ[1,2]) -> X[2],             Re(ρ[1,3]) -> X[3],             Re(ρ[1,4]) -> X[4],
Re(ρ[2,2]) -> X[5],             Re(ρ[2,3]) -> X[6],             Re(ρ[2,4]) -> X[7],
Re(ρ[3,3]) -> X[8],             Re(ρ[3,4]) -> X[9],
Re(ρ[3,3]) -> X[10]
Im(ρ[1,1]) -> X[11+LC+LK],      Im(ρ[1,2]) -> X[12+LC+LK],      Im(ρ[1,3]) -> X[13+LC+LK],      Im(ρ[1,4]) -> X[14+LC+ LK],
Im(ρ[2,2]) -> X[15+LC+LK],      Im(ρ[2,3]) -> X[16+LC+LK],      Im(ρ[2,4]) -> X[17+LC+LK],
Im(ρ[3,3]) -> X[18+LC+LK],      Im(ρ[3,4]) -> X[19+LC+LK],
Im(ρ[4,4]) -> X[20+LC+LK]

C: [0,1], [0,2], ..., [0,L], [1,1], ..., [1,L], [2,2], ..., [2,L], [3,3], ...., [L-1,L-1], [L-1,L], [L,L] Re -> [11, ..., 10 + LC] and Im -> [21 + LC + LK, ..., 20 + 2 LC + LK]

Re(C[i,j]) -> X[10+Q[i,j]]
            Re(C[0,1]) -> X[11],    Re(C[0,2]) -> X[12],    ...,
            Re(C[1,1]) -> X[11+L],  ...,    Re(C[L,L]) -> X[10+LC]
Im(C[i,j]) -> X[10+Q[i,j] + LIm]
            Im(C[0,1]) -> X[11+LIm],      Im(C[0,2]) -> X[12+LIm],     ...,
            Im(C[1,1]) -> X[11+L+LIm],    ...,    Im(C[L,L]) -> X[10+LC+LIm]

K: [1,2], ..., [1,L], [2,3], ..., [2,L], [3,4], ...., [3,L], [4,5], ..., [L-2,L-1], [L-2,L], [L-1,L] Re -> [11 + LC, ..., LIm] and Im -> [21 + 2 LC + LK, ..., 20 + 2 LC + 2 LK]

Re(K[i,j]) -> X[10+LC+P[i,j]]
            Re(K[1,2]) -> X[11+LC],     Re(K[1,3]) -> X[12+LC],     ...,
            Re(K[2,3]) -> X[10+LC+L),   ...,    Re(K[L-1,L]) ->  X[LIm]
Im(K[i,j]) -> X[10+LC+P[i,j]+LIm]
            Im(K[1,2]) -> X[11+LC+LIm],     Im(K[1,3]) -> X[12+LC+LIm],     ...,
            Im(K[2,3]) -> X[10+LC+L+LIm),   ...,    Im(K[L-1,L]) ->  X[2 LIm]

MFDecoupling.setup_calculation — Method

setup_calculation(fp1::String, fp2::String, L::Int; mode::Symbol=:real)
setup_calculation(X0::Vector, L::Int; mode=:real)

Sets up the calculation. Returns start vector X0 and function rhs for the right hand side of differential equation. Mode can either be :real or :complex. The former uses a real vector representation (with twice as many vector entries) and is the (much fast) default.

It may be useful for performance, to read inputs only a single time with read_inputs, which returns X0.